requirements.txt
# Data Collection
selenium>=4.15.0
webdriver-manager>=4.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
pytz>=2023.3

# FRED API
fredapi>=0.5.1

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.2
click>=8.1.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.26.0
